requirements.txt
pydub
pyannote.audio
torch
transformers
torchvision
speechRecognition
spacy